Move MIME-type detection from class CFileUtils to right class CMime
