+++ xbmc/utils/Mime.h
@@ -33,6 +33,23 @@ public:
   static std::string GetMimeType(const CFileItem &item);
   static std::string GetMimeType(const CURL &url, bool lookup = true);
 
+  enum EFileType
+  {
+    FileTypeUnknown = 0,
+    FileTypeHtml,
+    FileTypeXml,
+    FileTypePlainText,
+    FileTypeZip,
+    FileTypeGZip,
+    FileTypeRar,
+    FileTypeBmp,
+    FileTypeGif,
+    FileTypePng,
+    FileTypeJpeg,
+  };
+  static EFileType GetFileTypeFromMime(const std::string& mimeType);
+  static bool parseMimeType(const std::string& mimeType, std::string& type, std::string& subtype);
+
 private:
   static std::map<std::string, std::string> m_mimetypes;
 };
